minor commit: restructure
this commit just moves some parts that were in main.go into new files,
to better describe their function. For example messageCreate() is now in
bot.go, and the Cache variable is defined in bot.go as well. common.go
contains the simple, misc functions that clean input, or log messages.

go.sum is also updated, but that is a thing the go toolkit did.