simplify cmd/search
Only needed to return the url of the xkcd, now only returns the url of
the xkcd.

This may be useful for hands-on testing the bot, so we don't have to
query Google every time even when we're just testing searches.

Need to find out a solution for that, or bot will keep getting blocked
by Google :(